+++ src/request_queue.js
@@ -26,6 +26,15 @@ export const API_PROCESSED_REQUESTS_DELAY_MILLIS = 10 * 1000;
 // How many times we try to get queue head with queueModifiedAt older than API_PROCESSED_REQUESTS_DELAY_MILLIS.
 export const MAX_QUERIES_FOR_CONSISTENCY = 6;
 
+// This number must be large enough so that processing of all these requests cannot be done in
+// a time lower than expected maximum latency of DynamoDB, but low enough not to waste too much memory.
+const RECENTLY_HANDLED_CACHE_SIZE = 1000;
+
+// Indicates how long it usually takes for the underlying storage to propagate all writes
+// to be available to subsequent reads.
+export const STORAGE_CONSISTENCY_DELAY_MILLIS = 3000;
+
+
 const writeFilePromised = promisify(fs.writeFile);
 const readdirPromised = promisify(fs.readdir);
 const readFilePromised = promisify(fs.readFile);
@@ -42,20 +51,17 @@ const queuesCache = new LruCache({ maxLength: MAX_OPENED_QUEUES }); // Open queu
  */
 const validateAddRequestParams = (request, opts) => {
     checkParamOrThrow(request, 'request', 'Object');
-
-    if (!(request instanceof Request)) {
-        request = new Request(request);
-    }
-
     checkParamOrThrow(opts, 'opts', 'Object');
+
+    const newRequest = request instanceof Request ? request : new Request(request);
 
     const { forefront = false } = opts;
 
     checkParamOrThrow(forefront, 'opts.forefront', 'Boolean');
 
-    if (request.id) throw new Error('Request has already "id" so it cannot be added to the queue!');
+    if (request.id) throw new Error('Request already has the "id" field set so it cannot be added to the queue!');
 
-    return { forefront, request };
+    return { forefront, newRequest };
 };
 
 /**
@@ -109,9 +115,7 @@ const getRequestId = (uniqueKey) => {
         .digest('base64')
         .replace(/(\+|\/|=)/g, '');
 
-    return str.length > 15
-        ? str.substr(0, 15)
-        : str;
+    return str.substr(0, 15);
 };
 
 /**
@@ -123,7 +127,7 @@ const getRequestId = (uniqueKey) => {
  * @property {Boolean} wasAlreadyPresent Indicates if request was already present in the queue.
  * @property {Boolean} wasAlreadyHandled Indicates if request was already marked as handled.
  * @property {String} requestId The ID of the added request
- * @property {Request} request The original `Request` object passed to the `RequestQueue` function.
+ * @property {Request|Object} request The original `Request` object passed to the `RequestQueue` function.
  */
 
 /**
@@ -197,24 +201,43 @@ export class RequestQueue {
         checkParamOrThrow(queueName, 'queueName', 'Maybe String');
         checkParamOrThrow(clientKey, 'clientKey', 'String');
 
-        if (!clientKey) throw new Error('Parameter "clientKey" must be nonempty string!');
+        if (!clientKey) throw new Error('Parameter "clientKey" must be a non-empty string!');
 
         this.clientKey = clientKey;
         this.queueId = queueId;
         this.queueName = queueName;
+
+        // Contains a cached list of request IDs from the head of the queue,
+        // as obtained in the last query. Both key and value is the request ID.
         this.queueHeadDict = new ListDictionary();
-        this.requestIdsInProgress = {};
-        this.inProgressCount = 0;
         this.queryQueueHeadPromise = null;
+
+        // A set of all request IDs that are currently being handled,
+        // i.e. which were returned by fetchNextRequest() but not markRequestHandled()
+        this.inProgress = new Set();
+
+        // Contains a list of recently handled requests. It is used to avoid inconsistencies
+        // caused by delays in the underlying DynamoDB storage.
+        // Keys are request IDs, values are true.
+        this.recentlyHandled = new LruCache({ maxLength: RECENTLY_HANDLED_CACHE_SIZE });
 
         // We can trust these numbers only in a case that queue is used by a single client.
         // This information is returned by getHead() under the hadMultipleClients property.
         this.assumedTotalCount = 0;
         this.assumedHandledCount = 0;
 
-        // Caching requests to avoid duplicite addRequest() calls.
+        // Caching requests to avoid redundant addRequest() calls.
         // Key is computed using getRequestId() and value is { id, isHandled }.
+        // TODO: We could extend the caching to improve performance
+        //       of other operations such as fetchNextRequest().
         this.requestsCache = new LruCache({ maxLength: MAX_CACHED_REQUESTS });
+    }
+
+    /**
+     * @ignore
+     */
+    inProgressCount() {
+        return this.inProgress.size;
     }
 
     /**
@@ -224,187 +247,240 @@ export class RequestQueue {
      * it will not be updated. You can find out whether this happened from the resulting
      * {@link QueueOperationInfo} object.
      *
-     * @param {Request|Object} request {@link Request} object, or an object to construct a `Request` instance from.
+     * @param {Request|Object} request {@link Request} object or vanilla object with request data.
+     * Note that the function sets the `uniqueKey` and `id` fields to the passed object.
      * @param {Object} [options]
      * @param {Boolean} [options.forefront=false] If `true`, the request will be added to the foremost position in the queue.
      * @return {QueueOperationInfo}
      */
-    addRequest(request, options = {}) {
-        const { forefront, request: newRequest } = validateAddRequestParams(request, options);
+    async addRequest(request, options = {}) {
+        const { newRequest, forefront } = validateAddRequestParams(request, options);
 
-        if (newRequest) {
-            request = newRequest;
-        }
+        request.uniqueKey = newRequest.uniqueKey;
 
-        const cacheKey = getRequestId(request.uniqueKey);
+        const cacheKey = getRequestId(newRequest.uniqueKey);
         const cachedInfo = this.requestsCache.get(cacheKey);
 
         if (cachedInfo) {
-            return Promise.resolve({
+            request.id = cachedInfo.id;
+            return {
                 wasAlreadyPresent: true,
                 // We may assume that if request is in local cache then also the information if the
                 // request was already handled is there because just one client should be using one queue.
                 wasAlreadyHandled: cachedInfo.isHandled,
                 requestId: cachedInfo.id,
-                // TODO: Why not set request.id to cachedInfo.id???
                 request,
-            });
+            };
         }
 
-        return requestQueues
-            .addRequest({
-                request,
-                queueId: this.queueId,
-                forefront,
-                clientKey: this.clientKey,
-            })
-            .then((queueOperationInfo) => {
-                const { requestId, wasAlreadyHandled, wasAlreadyPresent } = queueOperationInfo;
+        const queueOperationInfo = await requestQueues.addRequest({
+            request: newRequest,
+            queueId: this.queueId,
+            forefront,
+            clientKey: this.clientKey,
+        });
 
-                this._cacheRequest(cacheKey, queueOperationInfo);
+        const { requestId, wasAlreadyPresent } = queueOperationInfo;
 
-                if (forefront && !this.requestIdsInProgress[requestId] && !wasAlreadyHandled) {
-                    this.queueHeadDict.add(requestId, requestId, true);
-                }
+        this._cacheRequest(cacheKey, queueOperationInfo);
 
-                if (!wasAlreadyPresent) {
-                    this.assumedTotalCount++;
-                }
+        if (!wasAlreadyPresent && !this.inProgress.has(requestId) && !this.recentlyHandled.get(requestId)) {
+            this.assumedTotalCount++;
 
-                // TODO: Why not set request.id to cachedInfo.id???
-                queueOperationInfo.request = request;
+            // Performance optimization: add request straight to head if possible
+            this._maybeAddRequestToQueueHead(requestId, forefront);
+        }
 
-                return queueOperationInfo;
-            });
+        request.id = requestId;
+        queueOperationInfo.request = request;
+
+        return queueOperationInfo;
     }
 
     /**
      * Gets the request from the queue specified by ID.
      *
-     * @param {String} requestId Request ID
+     * @param {String} requestId ID of the request.
      * @return {Promise<Request>} Returns the request object, or `null` if it was not found.
      */
     async getRequest(requestId) {
         validateGetRequestParams(requestId);
 
         // TODO: Could we also use requestsCache here? It would be consistent with addRequest()
+        // Downside is that it wouldn't reflect changes from outside...
         const obj = await requestQueues.getRequest({
             requestId,
             queueId: this.queueId,
         });
 
-        return obj ? new Request(obj) : obj;
+        return obj ? new Request(obj) : null;
     }
 
     /**
-     * Returns next request in the queue to be processed.
+     * Returns a next request in the queue to be processed, or `null` if there are no more pending requests.
      *
-     * @returns {Promise<Request>} Returns the request object, or `null` if there are no more pending requests.
+     * Once you successfully finish processing of the request, you need to call {@link RequestQueue.markRequestHandled}
+     * to mark the request as handled in the queue. If there was some error in processing the request,
+     * call {@link RequestQueue.reclaimRequest} instead, so that the queue will give the request to some other consumer
+     * in another call to the `fetchNextRequest` function.
+     *
+     * Note that the `null` return value doesn't mean the queue processing finished,
+     * it means there are currently no pending requests.
+     * To check whether all requests in queue were finished, use {@link RequestQueue.isFinished} instead.
+     *
+     * @returns {Promise<Request>}
+     * Returns the request object or `null` if there are no more pending requests.
      */
-    fetchNextRequest() {
-        return this
-            ._ensureHeadIsNonEmpty()
-            .then(() => {
-                const nextId = this.queueHeadDict.removeFirst();
+    async fetchNextRequest() {
+        await this._ensureHeadIsNonEmpty();
 
-                // We are likely done at this point.
-                if (!nextId) return null;
+        const nextRequestId = this.queueHeadDict.removeFirst();
 
-                this._addToInProgress(nextId);
+        // We are likely done at this point.
+        if (!nextRequestId) return null;
 
-                return this
-                    .getRequest(nextId)
-                    .then((request) => {
-                        // TODO: What happens when getRequest() fails? Shouldn't we run the code below too???
-                        // We need to handle this situation because request may not be available
-                        // immediately after adding to the queue.
-                        if (!request) {
-                            this._removeFromInProgress(nextId);
-                            this.queueHeadDict.add(nextId, nextId, false);
-                        }
-
-                        return request;
-                    });
+        // This should never happen, but...
+        if (this.inProgress.has(nextRequestId) || this.recentlyHandled.get(nextRequestId)) {
+            log.warning('Queue head returned a request that is already in progress?!', {
+                nextRequestId,
+                inProgress: this.inProgress.has(nextRequestId),
+                recentlyHandled: !!this.recentlyHandled.get(nextRequestId),
             });
+            return null;
+        }
+
+        this.inProgress.add(nextRequestId);
+
+        let request;
+        try {
+            request = await this.getRequest(nextRequestId);
+        } catch (e) {
+            // On error, remove the request from in progress, otherwise it would be there forever
+            this.inProgress.delete(nextRequestId);
+            throw e;
+        }
+
+        // NOTE: It can happen that the queue head index is inconsistent with the main queue table. This can occur in two situations:
+
+        // 1) Queue head index is ahead of the main table and the request is not present in the main table yet (i.e. getRequest() returned null).
+        //    In this case, keep the request marked as in progress for a short while,
+        //    so that isFinished() doesn't return true and _ensureHeadIsNonEmpty() doesn't not load the request
+        //    into the queueHeadDict straight again. After the interval expires, fetchNextRequest()
+        //    will try to fetch this request again, until it eventually appears in the main table.
+        if (!request) {
+            log.debug('Cannot find a request from the beginning of queue, will be retried later', { nextRequestId });
+            setTimeout(() => {
+                this.inProgress.delete(nextRequestId);
+            }, STORAGE_CONSISTENCY_DELAY_MILLIS);
+            return null;
+        }
+
+        // 2) Queue head index is behind the main table and the underlying request was already handled
+        //    (by some other client, since we keep the track of handled requests in recentlyHandled dictionary).
+        //    We just add the request to the recentlyHandled dictionary so that next call to _ensureHeadIsNonEmpty()
+        //    will not put the request again to queueHeadDict.
+        if (request.handledAt) {
+            log.debug('Request fetched from the beginning of queue was already handled', { nextRequestId });
+            this.recentlyHandled.add(nextRequestId, true);
+            return null;
+        }
+
+        return request;
     }
 
     /**
-     * Marks request handled after successful processing.
+     * Marks a request that was previously returned by the {@link RequestQueue.fetchNextRequest} function
+     * as handled after successful processing.
+     * Handled requests will never again be returned by the `fetchNextRequest` function.
      *
      * @param {Request} request
      * @return {Promise<QueueOperationInfo>}
      */
-    markRequestHandled(request) {
+    async markRequestHandled(request) {
         // TODO: This function should also support object instead of Apify.Request()
         validateMarkRequestHandledParams(request);
 
-        if (!this.requestIdsInProgress[request.id]) {
-            throw new Error(`Cannot mark request ${request.id} as handled as it is not in progress!`);
+        if (!this.inProgress.has(request.id)) {
+            throw new Error(`Cannot mark request ${request.id} as handled, because it is not in progress!`);
         }
 
         if (!request.handledAt) request.handledAt = new Date();
 
-        return requestQueues
-            .updateRequest({
-                request,
-                queueId: this.queueId,
-                clientKey: this.clientKey,
-            })
-            .then((queueOperationInfo) => {
-                this._removeFromInProgress(request.id);
-                this._cacheRequest(getRequestId(request.uniqueKey), queueOperationInfo);
+        const queueOperationInfo = await requestQueues.updateRequest({
+            request,
+            queueId: this.queueId,
+            clientKey: this.clientKey,
+        });
 
-                if (!queueOperationInfo.wasAlreadyHandled) {
-                    this.assumedHandledCount++;
-                }
+        this.inProgress.delete(request.id);
+        this.recentlyHandled.add(request.id, true);
 
-                queueOperationInfo.request = request;
+        if (!queueOperationInfo.wasAlreadyHandled) {
+            this.assumedHandledCount++;
+        }
 
-                return queueOperationInfo;
-            });
+        this._cacheRequest(getRequestId(request.uniqueKey), queueOperationInfo);
+
+        queueOperationInfo.request = request;
+
+        return queueOperationInfo;
     }
 
     /**
-     * Reclaims failed request back to the queue, so that it can be processed later again.
+     * Reclaims a failed request back to the queue, so that it can be returned for processed later again
+     * by another call to {@link RequestQueue.fetchNextRequest}.
      * The request record in the queue is updated using the provided `request` parameter.
-     * For example, this lets you store the number of retries for the request.
+     * For example, this lets you store the number of retries or error messages for the request.
      *
      * @param {Request} request
      * @param {Object} [options]
      * @param {Boolean} [options.forefront=false]
-     *   If `true` then requests get returned to the start of the queue
-     *   and to the back of the queue otherwise.
+     * If `true` then the request it placed to the beginning of the queue, so that it's returned
+     * in the next call to {@link RequestQueue.fetchNextRequest}. By default, it's put to the end of the queue.
      * @return {Promise<QueueOperationInfo>}
      */
-    reclaimRequest(request, options = {}) {
+    async reclaimRequest(request, options = {}) {
+        // TODO: This function should also support object instead of Apify.Request()
         const { forefront } = validateReclaimRequestParams(request, options);
 
-        // TODO: This function should also support object instead of Apify.Request()
+        if (!this.inProgress.has(request.id)) {
+            throw new Error(`Cannot reclaim request ${request.id}, because it is not in progress!`);
+        }
+
         // TODO: If request hasn't been changed since the last getRequest(),
         // we don't need to call updateRequest() and thus improve performance.
-        return requestQueues
-            .updateRequest({
-                request,
-                queueId: this.queueId,
-                forefront,
-                clientKey: this.clientKey,
-            })
-            .then((queueOperationInfo) => {
-                this._removeFromInProgress(request.id);
-                this._cacheRequest(getRequestId(request.uniqueKey), queueOperationInfo);
 
-                if (forefront) this.queueHeadDict.add(request.id, request.id, true);
+        const queueOperationInfo = await requestQueues.updateRequest({
+            request,
+            queueId: this.queueId,
+            forefront,
+            clientKey: this.clientKey,
+        });
 
-                queueOperationInfo.request = request;
+        this._cacheRequest(getRequestId(request.uniqueKey), queueOperationInfo);
+        queueOperationInfo.request = request;
 
-                return queueOperationInfo;
-            });
+        // Wait a little to increase a chance that the next call to fetchNextRequest() will return the request with updated data.
+        // This is to compensate for the limitation of DynamoDB, where writes might not be immediately visible to subsequent reads.
+        setTimeout(() => {
+            if (!this.inProgress.has(request.id)) {
+                log.warning('The request is no longer marked as in progress in the queue?!', { requestId: request.id });
+                return;
+            }
+
+            this.inProgress.delete(request.id);
+
+            // Performance optimization: add request straight to head if possible
+            this._maybeAddRequestToQueueHead(request.id, forefront);
+        }, STORAGE_CONSISTENCY_DELAY_MILLIS);
+
+        return queueOperationInfo;
     }
 
     /**
-     * Resolves to `true` if the next call to {@link RequestQueue#fetchNextRequest} would return `null`, otherwise it resolves to `false`.
+     * Resolves to `true` if the next call to {@link RequestQueue.fetchNextRequest} would return `null`, otherwise it resolves to `false`.
      * Note that even if the queue is empty, there might be some pending requests currently being processed.
-     * If you need to ensure that there is no activity in the queue, use {@link RequestQueue#isFinished}.
+     * If you need to ensure that there is no activity in the queue, use {@link RequestQueue.isFinished}.
      *
      * @returns {Promise<Boolean>}
      */
@@ -415,15 +491,17 @@ export class RequestQueue {
 
     /**
      * Resolves to `true` if all requests were already handled and there are no more left.
-     * Due to the nature of distributed storage systems,
-     * the function might occasionally return a false negative, but it will never return a false positive.
+     * Due to the nature of distributed storage used by the queue,
+     * the function might occasionally return a false negative,
+     * but it will never return a false positive.
      *
      * @returns {Promise<Boolean>}
      */
     async isFinished() {
-        if (this.inProgressCount > 0 || this.queueHeadDict.length() > 0) return false;
+        if (this.queueHeadDict.length() > 0 || this.inProgressCount() > 0) return false;
+
         const isHeadConsistent = await this._ensureHeadIsNonEmpty(true);
-        return isHeadConsistent && this.inProgressCount === 0 && this.queueHeadDict.length() === 0;
+        return isHeadConsistent && this.queueHeadDict.length() === 0 && this.inProgressCount() === 0;
     }
 
     /**
@@ -444,34 +522,7 @@ export class RequestQueue {
     }
 
     /**
-     * @ignore
-     */
-    _addToInProgress(requestId) {
-        checkParamOrThrow(requestId, 'requestId', 'String');
-
-        // Is already there.
-        if (this.requestIdsInProgress[requestId]) return;
-
-        this.requestIdsInProgress[requestId] = requestId;
-        this.inProgressCount++;
-    }
-
-    /**
-     * @ignore
-     */
-    _removeFromInProgress(requestId) {
-        checkParamOrThrow(requestId, 'requestId', 'String');
-
-        // Is already removed.
-        if (!this.requestIdsInProgress[requestId]) return;
-
-        delete this.requestIdsInProgress[requestId];
-        this.inProgressCount--;
-    }
-
-    /**
-     * We always request more items than is in progress to ensure that something
-     * falls into head.
+     * We always request more items than is in progress to ensure that something falls into head.
      *
      * @param {Boolean} [ensureConsistency=false] If true then query for queue head is retried until queueModifiedAt
      *   is older than queryStartedAt by at least API_PROCESSED_REQUESTS_DELAY_MILLIS to ensure that queue
@@ -483,7 +534,7 @@ export class RequestQueue {
      */
     async _ensureHeadIsNonEmpty(
         ensureConsistency = false,
-        limit = Math.max(this.inProgressCount * QUERY_HEAD_BUFFER, QUERY_HEAD_MIN_LENGTH),
+        limit = Math.max(this.inProgressCount() * QUERY_HEAD_BUFFER, QUERY_HEAD_MIN_LENGTH),
         iteration = 0,
     ) {
         checkParamOrThrow(ensureConsistency, 'ensureConsistency', 'Boolean');
@@ -491,7 +542,7 @@ export class RequestQueue {
         checkParamOrThrow(iteration, 'iteration', 'Number');
 
         // If is nonempty resolve immediately.
-        if (this.queueHeadDict.length()) return true;
+        if (this.queueHeadDict.length() > 0) return true;
 
         if (!this.queryQueueHeadPromise) {
             const queryStartedAt = new Date();
@@ -503,18 +554,19 @@ export class RequestQueue {
                     clientKey: this.clientKey,
                 })
                 .then(({ items, queueModifiedAt, hadMultipleClients }) => {
-                    items.forEach(({ id, uniqueKey }) => {
-                        if (!this.requestIdsInProgress[id]) {
-                            this.queueHeadDict.add(id, id, false);
-                            this._cacheRequest(getRequestId(uniqueKey), { requestId: id, wasAlreadyHandled: false });
-                        }
+                    items.forEach(({ id: requestId, uniqueKey }) => {
+                        // Queue head index might be behind the main table, so ensure we don't recycle requests
+                        if (this.inProgress.has(requestId) || this.recentlyHandled.get(requestId)) return;
+
+                        this.queueHeadDict.add(requestId, requestId, false);
+                        this._cacheRequest(getRequestId(uniqueKey), { requestId, wasAlreadyHandled: false });
                     });
 
-                    // This is needed so that the next call can request the queue head again.
+                    // This is needed so that the next call to _ensureHeadIsNonEmpty() will fetch the queue head again.
                     this.queryQueueHeadPromise = null;
 
                     return {
-                        wasLimitReached: items.length === limit,
+                        wasLimitReached: items.length >= limit,
                         prevLimit: limit,
                         queueModifiedAt: new Date(queueModifiedAt),
                         queryStartedAt,
@@ -525,6 +577,8 @@ export class RequestQueue {
 
         const { queueModifiedAt, wasLimitReached, prevLimit, queryStartedAt, hadMultipleClients } = await this.queryQueueHeadPromise;
 
+        // TODO: I feel this code below can be greatly simplified...
+
         // If queue is still empty then one of the following holds:
         // - the other calls waiting for this promise already consumed all the returned requests
         // - the limit was too low and contained only requests in progress
@@ -532,10 +586,12 @@ export class RequestQueue {
         // - the whole queue was processed and we are done
 
         // If limit was not reached in the call then there are no more requests to be returned.
-        const shouldRepeatWithHigherLimit = this.queueHeadDict.length() === 0 && wasLimitReached && prevLimit < REQUEST_QUEUE_HEAD_MAX_LIMIT;
         if (prevLimit >= REQUEST_QUEUE_HEAD_MAX_LIMIT) {
             log.warning(`RequestQueue: Reached the maximum number of requests in progress: ${REQUEST_QUEUE_HEAD_MAX_LIMIT}.`);
         }
+        const shouldRepeatWithHigherLimit = this.queueHeadDict.length() === 0
+            && wasLimitReached
+            && prevLimit < REQUEST_QUEUE_HEAD_MAX_LIMIT;
 
         // If ensureConsistency=true then we must ensure that either:
         // - queueModifiedAt is older than queryStartedAt by at least API_PROCESSED_REQUESTS_DELAY_MILLIS
@@ -567,29 +623,41 @@ export class RequestQueue {
     }
 
     /**
+     * Adds a request straight to the queueHeadDict, to improve performance.
+     * @private
+     */
+    _maybeAddRequestToQueueHead(requestId, forefront) {
+        if (forefront) {
+            this.queueHeadDict.add(requestId, requestId, true);
+        } else if (this.assumedTotalCount < QUERY_HEAD_MIN_LENGTH) {
+            this.queueHeadDict.add(requestId, requestId, false);
+        }
+    }
+
+    /**
      * Removes the queue either from the Apify Cloud storage or from the local directory,
      * depending on the mode of operation.
      *
      * @return {Promise}
      */
-    delete() {
-        return requestQueues
-            .deleteQueue({
-                queueId: this.queueId,
-            })
-            .then(() => {
-                queuesCache.remove(this.queueId);
-                if (this.queueName) queuesCache.remove(this.queueName);
-            });
+    async delete() {
+        await requestQueues.deleteQueue({
+            queueId: this.queueId,
+        });
+
+        queuesCache.remove(this.queueId);
+        if (this.queueName) queuesCache.remove(this.queueName);
     }
 
     /**
      * Returns the number of handled requests.
      *
      * @return {Promise<number>}
+     * @deprecated
      */
     async handledCount() {
-        // TODO: We should deprectate this function in favor of getInfo()
+        log.deprecated('RequestQueue.handledCount() is deprecated, please use RequestQueue.getInfo() instead.');
+
         const queueInfo = await requestQueues.getQueue({ queueId: this.queueId });
         return queueInfo.handledRequestCount;
     }
@@ -640,6 +708,9 @@ const filePathToQueueOrderNo = (filepath) => {
 
 /**
  * Local directory-based implementation of the `RequestQueue` class.
+ * TODO: We should implement this class using the RequestQueueRemote, just replace
+ * the underlying API calls with their emulation on filesystem. That will bring
+ * all the goodies such as caching and will enable better and more consistent testing
  *
  * @ignore
  */
@@ -739,12 +810,10 @@ export class RequestQueueLocal {
             });
     }
 
-    addRequest(request, opts = {}) {
-        const { forefront, request: newRequest } = validateAddRequestParams(request, opts);
+    async addRequest(request, opts = {}) {
+        const { newRequest, forefront } = validateAddRequestParams(request, opts);
 
-        if (newRequest) {
-            request = newRequest;
-        }
+        request.uniqueKey = newRequest.uniqueKey;
 
         return this.initializationPromise
             .then(() => {
@@ -752,8 +821,9 @@ export class RequestQueueLocal {
 
                 // Add ID as server does.
                 // TODO: This way of cloning doesn't preserve Dates!
-                const requestCopy = JSON.parse(JSON.stringify(request));
+                const requestCopy = JSON.parse(JSON.stringify(newRequest));
                 requestCopy.id = getRequestId(request.uniqueKey);
+                request.id = requestCopy.id;
 
                 this._updateMetadata(true);
 
@@ -844,7 +914,7 @@ export class RequestQueueLocal {
                 const dest = this._getFilePath(queueOrderNo, true);
 
                 if (!this.queueOrderNoInProgress[queueOrderNo]) {
-                    throw new Error(`Cannot mark request ${request.id} handled request that is not in progress!`);
+                    throw new Error(`Cannot mark request ${request.id} as handled, because it is not in progress!`);
                 }
 
                 if (!request.handledAt) request.handledAt = new Date();
@@ -882,7 +952,7 @@ export class RequestQueueLocal {
                 const dest = this._getFilePath(newQueueOrderNo);
 
                 if (!this.queueOrderNoInProgress[oldQueueOrderNo]) {
-                    throw new Error(`Cannot reclaim request ${request.id} that is not in progress!`);
+                    throw new Error(`Cannot reclaim request ${request.id}, because it is not in progress!`);
                 }
 
                 this.requestIdToQueueOrderNo[request.id] = newQueueOrderNo;